1.0.0.69
New: Equipment view updated to show current sentient gem and filigrees
---This will eventually become interactive to set filigrees for gem and lesser artifacts
---How To Guide updated
Fix: The XML reader will now report the actual error message rather than "Unspecified error" on a file read failure.

Update 41.4 Pre-order changes
New: Racial completionist feat now requires 3 Tiefling past lives to become active
New: New special feat "Inherent Universal Action Point" added
---Gives you an AP which can only be spent in a universal (non-class, non-racial) tree
---All universal trees are now marked internally as universal

+++ DDOCP/EnhancementTreeItem.cpp
@@ -205,8 +205,8 @@ bool EnhancementTreeItem::CanTrain(
     // verify and enhancements we are dependent on have enough trained ranks also
     canTrain &= m_RequirementsToTrain.CanTrainEnhancement(charData, trainedRanks);
     // must have enough action points to buy it
-    size_t availableAP = charData.AvailableActionPoints(treeName, type);
-    canTrain &= (availableAP >= Cost());
+    int availableAP = charData.AvailableActionPoints(treeName, type);
+    canTrain &= (availableAP >= (int)Cost());
     return canTrain;
 }
